pyproject.toml


[project]
name = "pangeo-forge-ndpyramid"
description = "Pangeo Forge extension libary to generate multi-scale pyramids through pangeo-forge-recipes."
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "CarbonPlan", email = "tech@carbonplan.org"}]
requires-python = ">=3.9"
dynamic = ["version"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering",
]

dependencies = [
    "pangeo-forge-recipes",
    "ndpyramid@git+https://github.com/carbonplan/ndpyramid@resample-blocks",
    "dask",
    "pyresample",
    "rioxarray"
]

[project.optional-dependencies]
test = [
    "pre-commit",
    "pytest-mypy",
    "pytest-cov",
    "pytest",
    "ruff",
    "rioxarray",
    "pooch",
    "xarray-datatree"

]
dynamic = ["version"]


[build-system]
requires = [
    "setuptools>=64",
    "setuptools-scm[toml]>=6.2"
]
build-backend = "setuptools.build_meta"


[tool.setuptools_scm]
write_to = "pangeo_forge_ndpyramid/_version.py"
write_to_template = '''
__version__ = "{version}"
'''

[tool.setuptools.packages.find]
exclude = ["docs", "tests", "tests.*", "docs.*"]




[mypy]
files = "pangeo-forge-ndpyramid/**/*.py"
show_error_codes = true

[tool.ruff]
# Same as Black.
line-length = 88
indent-width = 4
target-version = "py39"

exclude = [
    ".eggs"]

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F", "I"]
per-file-ignores = {}

# E402: module level import not at top of file
# E731: do not assign a lambda expression, use a def
# W503: line break before binary operator - not implimeted by ruff due to conflict with PEP8.

ignore = ["E402", "E731"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"
# Indent with spaces, rather than tabs.
indent-style = "space"
# Respect magic trailing commas.
skip-magic-trailing-comma = false
# Automatically detect the appropriate line ending.
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["pangeo-forge-ndpyramid"]
